requirements.txt
tk==0.1.0
pygame==2.1.3
PyOpenGL==3.1.6
